Check safety of Immutable
