+++ main/src/ca/uwaterloo/flix/language/errors/SafetyError.scala
@@ -530,4 +530,34 @@ object SafetyError {
          |""".stripMargin
     })
   }
+
+  /**
+    * Unable to derive Immutable error
+    *
+    * @param sym the enum for which we're trying to derive Immutable
+    * @param loc the location where the error occurred.
+    */
+  case class ImmutableError(tpe: Type, loc: SourceLocation) extends SafetyError {
+    def summary: String = s"Cannot derive Immutable for ${tpe}"
+
+    def message(formatter: Formatter): String = {
+      import formatter._
+      s"""${line(kind, source.name)}
+        |>> Cannot derive 'Immutable' for type ${red(tpe.toString)}
+        |
+        |Because it takes a type parameter of kind 'Region'.
+        |
+        |${code(loc, "unable to derive Immutable.")}
+        |
+        |""".stripMargin
+    }
+
+    def explain(formatter: Formatter): Option[String] = Some(
+      s"""
+        |An example of a type parameter of kind 'Region':
+        |
+        |enum MyEnum[r: Region] { ... }
+        |""".stripMargin
+    )
+  }
 }
