+++ main/src/ca/uwaterloo/flix/language/phase/Safety.scala
@@ -10,10 +10,11 @@ import ca.uwaterloo.flix.language.ast.{Kind, RigidityEnv, SourceLocation, Symbol
 import ca.uwaterloo.flix.language.errors.SafetyError
 import ca.uwaterloo.flix.language.errors.SafetyError._
 import ca.uwaterloo.flix.language.phase.unification.Unification
-import ca.uwaterloo.flix.util.Validation
+import ca.uwaterloo.flix.util.{Validation, InternalCompilerException}
 import ca.uwaterloo.flix.util.Validation._
 
 import scala.annotation.tailrec
+import ca.uwaterloo.flix.language.errors.DerivationError
 
 /**
   * Performs safety and well-formedness checks on:
@@ -32,9 +33,13 @@ object Safety {
     //
     // Collect all errors.
     //
-    val errors = root.defs.flatMap {
+    val defErrors = root.defs.flatMap {
       case (_, defn) => visitDef(defn)
     }
+
+    val instanceErrors = visitImmutable(root)
+
+    val errors = defErrors ++ instanceErrors
 
     //
     // Check if any errors were detected.
@@ -43,6 +48,22 @@ object Safety {
       root.toSuccess
     else
       Validation.Failure(errors.to(LazyList))
+  }
+
+  /**
+    * Checks that no type parameters for types that implement `Immutable` of kind `Region`
+    */
+  private def visitImmutable(root: Root)(implicit flix: Flix): List[CompilationMessage] = {
+
+    val immutableClass = new Symbol.ClassSym(Nil, "Immutable", SourceLocation.Unknown)
+
+    root.instances.getOrElse(immutableClass, Nil) flatMap {
+      case Instance(_, _, _, _, tpe, _, _, _, loc) =>
+        if (tpe.typeArguments.exists(_.kind == Kind.Bool))
+          List(SafetyError.ImmutableError(tpe, loc))
+        else
+          Nil
+    }
   }
 
   /**
